+++ src/lufah/__init__.py
@@ -1,6 +1,6 @@
 """lufah"""
 
-__version__ = "0.9.1"
+__version__ = "0.10.0"
 __all__ = ["COMMAND_FINISH", "COMMAND_FOLD", "COMMAND_PAUSE", "FahClient"]
 
 from .const import COMMAND_FINISH, COMMAND_FOLD, COMMAND_PAUSE